Do tasks

+++ product/models.py
@@ -5,7 +5,7 @@ from product.querysets import ProductQuerySet
 
 class Product(models.Model):
     name = models.CharField(max_length=100)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
     description = models.TextField()
     stock = models.IntegerField(default=0)
